+++ test/algorithms/physical_design/hexagonalization.cpp
@@ -26,23 +26,58 @@ using namespace fiction;
 template <typename Lyt, typename Ntk>
 void check_mapping_equiv(const Ntk& ntk)
 {
-    const auto             layout = orthogonal<Lyt>(ntk, {});
-    hexagonalization_stats stats{};
-    const auto             hex_layout = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(layout, &stats);
+    const auto layout = orthogonal<Lyt>(ntk, {});
+
+    hexagonalization_stats  stats{};
+    hexagonalization_params params{};
+    const auto              hex_layout = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(layout, params, &stats);
+
+    params.place_inputs_in_top_row = true;
+    const auto hex_layout_top_pis  = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(layout, params, &stats);
 
     check_eq(ntk, layout);
     check_eq(ntk, hex_layout);
+    check_eq(ntk, hex_layout_top_pis);
     check_eq(layout, hex_layout);
+    check_eq(layout, hex_layout_top_pis);
+
+    hex_layout_top_pis.foreach_pi([&](const auto& gate) { CHECK(hex_layout_top_pis.get_tile(gate).y == 0); });
 }
 
 template <typename Lyt>
 void check_mapping_equiv_layout(const Lyt& lyt)
 {
-    hexagonalization_stats stats{};
-    const auto             hex_layout = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(lyt, &stats);
+    hexagonalization_stats  stats{};
+    hexagonalization_params params{};
+    const auto              hex_layout = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(lyt, params, &stats);
+
+    params.place_inputs_in_top_row = true;
+    const auto hex_layout_top_pis  = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(lyt, params, &stats);
 
     check_eq(lyt, hex_layout);
+    check_eq(lyt, hex_layout_top_pis);
+
     CHECK(lyt.get_layout_name() == hex_layout.get_layout_name());
+    CHECK(lyt.get_layout_name() == hex_layout_top_pis.get_layout_name());
+
+    hex_layout_top_pis.foreach_pi([&](const auto& gate) { CHECK(hex_layout_top_pis.get_tile(gate).y == 0); });
+}
+
+template <typename Lyt>
+void check_mapping_equiv_layout_with_planar_rerouting(const Lyt& lyt)
+{
+    hexagonalization_stats  stats{};
+    hexagonalization_params params{};
+
+    params.place_inputs_in_top_row         = true;
+    params.planar_routing_for_moved_inputs = true;
+    const auto hex_layout                  = hexagonalization<hex_even_row_gate_clk_lyt, Lyt>(lyt, params, &stats);
+
+    check_eq(lyt, hex_layout);
+
+    CHECK(lyt.get_layout_name() == hex_layout.get_layout_name());
+
+    hex_layout.foreach_pi([&](const auto& gate) { CHECK(hex_layout.get_tile(gate).y == 0); });
 }
 
 template <typename Lyt>
@@ -68,6 +103,13 @@ void check_mapping_equiv_all()
     check_mapping_equiv_layout(blueprints::or_not_gate_layout<cart_gate_clk_lyt>());
     check_mapping_equiv_layout(blueprints::crossing_layout<cart_gate_clk_lyt>());
     check_mapping_equiv_layout(blueprints::tautology_gate_layout<cart_gate_clk_lyt>());
+    check_mapping_equiv_layout(blueprints::ge_gt_le_lt_layout<cart_gate_clk_lyt>());
+
+    check_mapping_equiv_layout_with_planar_rerouting(blueprints::straight_wire_gate_layout<cart_gate_clk_lyt>());
+    check_mapping_equiv_layout_with_planar_rerouting(blueprints::or_not_gate_layout<cart_gate_clk_lyt>());
+    check_mapping_equiv_layout_with_planar_rerouting(blueprints::crossing_layout<cart_gate_clk_lyt>());
+    check_mapping_equiv_layout_with_planar_rerouting(blueprints::tautology_gate_layout<cart_gate_clk_lyt>());
+    check_mapping_equiv_layout_with_planar_rerouting(blueprints::ge_gt_le_lt_layout<cart_gate_clk_lyt>());
 }
 
 TEST_CASE("Layout equivalence", "[hexagonalization]")